workspace/gateway/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
httpx==0.24.1
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0